[Fix] 매매 수량을 고정 원화 거래대금 기준으로 변경
Order.quantity/AgentHolding.quantity/Stock.volume을 Long에서 BigDecimal(scale 8)로 변경하고,
고정 수량(10개) 대신 고정 거래대금(100,000원)을 현재가로 나눈 수량으로 매매하도록 바꾼다.
코인마다 단가 편차가 커서(BTC 약 9,300만원 vs DOGE 약 280원) 고정 개수 매매는 BTC 거래 한 번에
초기자본의 900배가 넘는 금액이 나가는 문제가 있었다.

+++ backend/src/main/java/com/agentarena/backend/domain/order/dto/OrderResponse.java
@@ -12,7 +12,7 @@ public record OrderResponse(
         Long stockId,
         String stockCode,
         OrderType type,
-        Long quantity,
+        BigDecimal quantity,
         BigDecimal price,
         LocalDateTime executedAt
 ) {
