+++ backend/src/main/java/com/agentarena/backend/domain/stock/dto/StockResponse.java
@@ -3,7 +3,7 @@ package com.agentarena.backend.domain.stock.dto;
 import com.agentarena.backend.domain.stock.Stock;
 import java.math.BigDecimal;
 
-public record StockResponse(Long id, String code, String name, BigDecimal currentPrice, Long volume) {
+public record StockResponse(Long id, String code, String name, BigDecimal currentPrice, BigDecimal volume) {
 
     public static StockResponse from(Stock stock) {
         return new StockResponse(
